Integrating final MSFT fixes
Now the last little changes have been implemented, had to modify how we got the user to message since we were previously getting the first user mentioned in the message, whereas we now match against all known users in the guild. This means users actually get messaged by, say, rolebot even if they're not in the channel you're tipping from. Of course you can't directly mention anyone not in the channel for multitip, so that's a non-issue in most use cases.

Also had to ensure certain function calls were sending the right type, because dynamic languages and lack of type safety.

That _should_ be absolutely everything needed to have it fully functional.