Fix Private Mode Usernames
So it turns out that if you try to print a userid of someone that isn't in that server, such as in a dm channel, it simply displays as invalid user (Even though userids are unique across the entirety of discord, got to love that logic). So now the private message instead uses the .username and .tag attributes to provide the normal userid, like so: @GenericUser#1234

While this may be a tad confusing to those using nicknames, it identifies the exact individual doing the tipping, it just might require a little digging.