Tipbot Fixed & Made Easier to Use
I've combined the tip help messages into one easy-to-read message to make it easier for people to access and use. If anyone can't understand the obviousness of the help, they can visit the tip FAQ page for a better understanding.

However, because I have no Bot testing server, I have left some code in this version that's no longer required in case removing anything breaks it,

It will be best to remove this at some point.