Updated tipbot.js (Fixed some messages)
Changed a few things with when a tip is sent or errors out. (Instead of just "I don't know how to tip that amount", it sounds a bit better and more accurate.)
Changed the "Wubba lubba dub dub" message to something a bit more... normal? 
Also changed some phrasing with messages in general. Nothing that will change the functionality, more just the general appearance of the messages.